+++ CHANGELOG.md
@@ -5,11 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Version 0.2.7 - 2025-09-09
+## Version 0.2.7 - 2025-09-xx
 
 ### Fixed
 
-- Reduce rate limit exceeded messages level to `warning` 
+- tbd
 
 ## Version 0.2.6 - 2025-08-04